requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
cryptography==41.0.7
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
Werkzeug==3.0.1
pypdf==6.1.2
PyMuPDF
reportlab
pylibdmtx
python-barcode
Pillow
gunicorn
